LPD-11253 Simplify Types into its separate file
